pyproject.toml
[tool.poetry]
name = "gai-tool"
version = "0.3.0"
description = "AI-Powered Automation Tool for Git 🚀"
authors = ["Daniel Ratmiroff <danielratmiroff@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
appdirs = "^1.4.4"
requests = "^2.32.3"
rich = "^13.9.4"
twine = "^6.1.0"
wheel = "^0.45.1"
zipp = "^3.21.0"
pick = "^2.4.0"
pytest = "^8.3.4"
pytest-cov = "^4.1.0"
groq = "^0.18.0"
build = "^1.2.2.post1"
colorama = "^0.4.6"
codecov = "^2.1.13"
huggingface-hub = "^0.29.1"
ollama = "^0.4.7"
langchain-ollama = "^0.2.3"
langchain-community = "^0.3.18"
langchain-core = "^0.3.37"
transformers = "^4.49.0"


[tool.poetry.group.dev.dependencies]
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gai = "gai_tool.main:main"
test = "tests.main:main"

[tool.poetry.urls]
homepage = "https://github.com/Danielratmiroff/gai"
repository = "https://github.com/Danielratmiroff/gai"
issues = "https://github.com/Danielratmiroff/gai/issues"
documentation = "https://github.com/Danielratmiroff/gai#readme"
